LayoutTests/imported/w3c/web-platform-tests/domxpath/fn-substring-before-expected.txt

^^^bar$$$bar$bar

FAIL substring-before() arguments depending on the context node assert_equals: expected "^^^" but got "^^^bar"
